Resize the compact player from its original size to double
Dragging the bottom right corner sizes the compact player anywhere from
the skin's own size to double, snapping to either when close; the menu
offers original, one and a half, and double size, and the size is kept.

At a whole number of screen pixels per skin pixel it draws pixel for
pixel as before. Between two, it draws at the larger into a cached layer
that the compositor scales down with bilinear filtering on its own
thread, so every skin pixel stays the same size and only its edges
soften. The smoothing is set on an outer panel alone, and the cache on
an inner one holding the player, so the sprites stay sharp inside the
layer and the layer redraws as the player changes.

The probe checks the in-between size in the real window, and the
capture tool can set the size, drag the corner and keep an earlier
frame to compare.

+++ src/Tuxflix.Core/Settings/AppSettings.cs
@@ -38,8 +38,8 @@ public sealed class ClassicPlayerSettings
     /// <summary>The <c>.wsz</c> file in use; none means the classic base skin.</summary>
     public string? Skin { get; set; }
 
-    /// <summary>Pixels per skin pixel: 2 ("double size") suits today's screens.</summary>
-    public int Scale { get; set; } = 2;
+    /// <summary>Screen pixels per skin pixel, 1 to 2: 2 ("double size") suits today's screens.</summary>
+    public double Scale { get; set; } = 2;
 
     public bool ShowEqualizer { get; set; } = true;
